Worked On Mover, Health, and Score Scritps
Did some work to draft up a few test scripts for Health, Score, and a script I call AutomaticMover for well, automatic movement. I'm kind of exhausted of coding for right now so I'll return in the morning to see if I can think up ways to make the score and health scripts work.